requirements.txt
click>=8.0.0
Pillow>=9.0.0
requests>=2.28.0
tqdm>=4.64.0
python-dotenv>=0.19.0
